Remove Connection and ConnectionCounterparty traits
